Review: 컴포넌트 분리, 유실된 이미지 연결
